conditional statement

+++ test.sh
@@ -1,13 +1,29 @@
+#Print number of cores
 core=$(sysctl -n hw.ncpu)
 echo $core
+
+#Print total CPU percentage (can be over 100%)
 s=$(ps -A -o %cpu | awk '{s+=$1} END {print s}')
-per=$(printf "%.0f" $s)
 Echo $s
+
+#Print total CPU percentage (can be over 100%) floored to the nearest whole number
+per=$(printf "%.0f" $s)
 echo $per
+
+#Print cpu percentage divided by number of cores
 cpu=$(($per/$core))
 echo $cpu
-NAME=$(whoami)
-echo $NAME
-mockingbird=$(top -F -R -o cpu -U $NAME -l 2 -n 2 -ncols 1 | sed -n '25p')
-echo $mockingbird
-kill $mockingbird
+
+if [$cpu -gt '80' ]
+then
+	#Print current user
+	NAME=$(whoami)
+	echo $NAME
+
+	#Print process with top cpu percentage
+	mockingbird=$(top -F -R -o cpu -U $NAME -l 2 -n 2 -ncols 1 | sed -n '25p')
+	echo $mockingbird
+
+	#Kills top process with CPU percentage
+	kill $mockingbird
+fi
